refactor(AutomaticLottoStorage): 구입 금액 기반으로 로또 발행 개수(=티켓) 반환 및 저장 위치 리팩토링
