+++ src/main/java/lotto/domain/AutomaticLottoStorage.java
@@ -1,18 +1,33 @@
 package lotto.domain;
 
 import java.util.List;
+import lotto.common.Constant;
+import lotto.common.ErrorMessage;
 
 public class AutomaticLottoStorage {
 
     private List<AutomaticLotto> automaticLottos;
-    private LottoStorage lottoStorage;
+    private final Integer totalPrice;
 
-    public AutomaticLottoStorage(final LottoStorage lottoStorage) {
-        this.automaticLottos = AutomaticLotto.from(lottoStorage.getTicketCount());
-        this.lottoStorage = lottoStorage;
+    public AutomaticLottoStorage(final Integer totalPrice) {
+        validate(totalPrice);
+
+        this.totalPrice = totalPrice;
+        this.automaticLottos = AutomaticLotto.from(getTicketCount());
     }
 
-    public static AutomaticLottoStorage of(final LottoStorage lottoStorage) {
-        return new AutomaticLottoStorage(lottoStorage);
+
+    public static AutomaticLottoStorage of(final Integer totalPrice) {
+        return new AutomaticLottoStorage(totalPrice);
+    }
+
+    private Integer getTicketCount() {
+        return this.totalPrice / Constant.LOTTO_TICKET_PRICE;
+    }
+
+    private void validate(final Integer totalPrice) {
+        if (totalPrice % Constant.LOTTO_TICKET_PRICE > 0) {
+            throw new IllegalArgumentException(ErrorMessage.LOTTO_PURCHASE_AMOUNT_INVALID_ERROR.getMessage());
+        }
     }
 }
